+++ include/fx/image.h
@@ -8,7 +8,7 @@
 typedef struct fx_image_s fx_image_t;
 
 int
-fx_image_init (fx_t *app, const char *url, size_t len, double x, double y, double width, double height, fx_image_t **result);
+fx_image_init (fx_t *app, const char *url, size_t len, float x, float y, float width, float height, fx_image_t **result);
 
 int
 fx_image_destroy (fx_image_t *image);
@@ -20,9 +20,9 @@ int
 fx_set_image_data (fx_image_t *image, void *data);
 
 int
-fx_get_image_bounds (fx_image_t *image, double *x, double *y, double *width, double *height);
+fx_get_image_bounds (fx_image_t *image, float *x, float *y, float *width, float *height);
 
 int
-fx_set_image_bounds (fx_image_t *image, double x, double y, double width, double height);
+fx_set_image_bounds (fx_image_t *image, float x, float y, float width, float height);
 
 #endif // FX_IMAGE_H
