+++ include/fx/text-input.h
@@ -10,7 +10,7 @@ typedef struct fx_text_input_s fx_text_input_t;
 typedef void (*fx_text_input_change_cb)(fx_text_input_t *text_input);
 
 int
-fx_text_input_init (fx_t *app, double x, double y, double width, double height, fx_text_input_t **result);
+fx_text_input_init (fx_t *app, float x, float y, float width, float height, fx_text_input_t **result);
 
 int
 fx_text_input_destroy (fx_text_input_t *text_input);
@@ -25,10 +25,10 @@ int
 fx_set_text_input_data (fx_text_input_t *text_input, void *data);
 
 int
-fx_get_text_input_bounds (fx_text_input_t *text_input, double *x, double *y, double *width, double *height);
+fx_get_text_input_bounds (fx_text_input_t *text_input, float *x, float *y, float *width, float *height);
 
 int
-fx_set_text_input_bounds (fx_text_input_t *text_input, double x, double y, double width, double height);
+fx_set_text_input_bounds (fx_text_input_t *text_input, float x, float y, float width, float height);
 
 char *
 fx_get_text_input_value (fx_text_input_t *text_input);
