+++ include/fx/text.h
@@ -9,7 +9,7 @@ typedef struct fx_text_s fx_text_t;
 typedef struct fx_text_span_s fx_text_span_t;
 
 int
-fx_text_init (fx_t *app, double x, double y, double width, double height, fx_text_t **result);
+fx_text_init (fx_t *app, float x, float y, float width, float height, fx_text_t **result);
 
 int
 fx_text_destroy (fx_text_t *text);
@@ -21,10 +21,10 @@ int
 fx_set_text_data (fx_text_t *text, void *data);
 
 int
-fx_get_text_bounds (fx_text_t *text, double *x, double *y, double *width, double *height);
+fx_get_text_bounds (fx_text_t *text, float *x, float *y, float *width, float *height);
 
 int
-fx_set_text_bounds (fx_text_t *text, double x, double y, double width, double height);
+fx_set_text_bounds (fx_text_t *text, float x, float y, float width, float height);
 
 int
 fx_append_text_span (fx_text_t *text, const char *value, size_t len, fx_text_span_t **result);
