+++ include/fx/view.h
@@ -6,7 +6,7 @@
 typedef struct fx_view_s fx_view_t;
 
 int
-fx_view_init (fx_t *app, double x, double y, double width, double height, fx_view_t **result);
+fx_view_init (fx_t *app, float x, float y, float width, float height, fx_view_t **result);
 
 int
 fx_view_destroy (fx_view_t *view);
@@ -18,9 +18,9 @@ int
 fx_set_view_data (fx_view_t *view, void *data);
 
 int
-fx_get_view_bounds (fx_view_t *view, double *x, double *y, double *width, double *height);
+fx_get_view_bounds (fx_view_t *view, float *x, float *y, float *width, float *height);
 
 int
-fx_set_view_bounds (fx_view_t *view, double x, double y, double width, double height);
+fx_set_view_bounds (fx_view_t *view, float x, float y, float width, float height);
 
 #endif // FX_VIEW_H
