+++ src/darwin/appkit/video.m
@@ -13,7 +13,7 @@
 @end
 
 int
-fx_video_init (fx_t *app, const char *url, size_t len, double x, double y, double width, double height, fx_video_t **result) {
+fx_video_init (fx_t *app, const char *url, size_t len, float x, float y, float width, float height, fx_video_t **result) {
   FXVideo *native_video = [[FXVideo alloc] init];
 
   native_video.frame = CGRectMake(x, y, width, height);
@@ -55,7 +55,7 @@ fx_set_video_data (fx_video_t *video, void *data) {
 }
 
 int
-fx_get_video_bounds (fx_video_t *video, double *x, double *y, double *width, double *height) {
+fx_get_video_bounds (fx_video_t *video, float *x, float *y, float *width, float *height) {
   NSRect frame = video->native_video.frame;
 
   if (x) *x = frame.origin.x;
@@ -67,7 +67,7 @@ fx_get_video_bounds (fx_video_t *video, double *x, double *y, double *width, dou
 }
 
 int
-fx_set_video_bounds (fx_video_t *video, double x, double y, double width, double height) {
+fx_set_video_bounds (fx_video_t *video, float x, float y, float width, float height) {
   video->native_video.frame = CGRectMake(x, y, width, height);
 
   return 0;
